General Improvements
Changing the names of some classes and methods in the Business, Web
Services and Database layer in Final Project.
Adding exception for layer in the Final Project and starting the
implementation of Web Services layer.
Adding the Widget to the Yamba Project and improvement in its Shared
Preferences for selecting the refreshing time.